requirements.txt
discord.py==1.3.4
asyncio==3.4.3
captcha==0.3
